+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # minify-html changelog
 
+## Pending
+
+- Disable treeshake annotations (e.g., `/*#__PURE__*/`) in minified JavaScript output as they are only useful for bundlers, not inline scripts.
+
 ## 0.18.1
 
 - Fix WASM builds.